[VIT-2879] Aggregation for activity during the day and workout session.

+++ VitalClient/src/main/java/io/tryvital/client/services/data/Summary.kt
@@ -32,9 +32,9 @@ data class WorkoutPayload(
     @Json(name = "sport")
     val sport: String,
     @Json(name = "calories")
-    val caloriesInKiloJules: Long,
+    val caloriesInKiloJules: Double,
     @Json(name = "distance")
-    val distanceInMeter: Long,
+    val distanceInMeter: Double,
     @Json(name = "heart_rate")
     val heartRate: List<QuantitySamplePayload>,
     @Json(name = "respiratory_rate")
